requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
schedule==1.2.0
requests==2.31.0
tzlocal==5.2
pytz==2023.3
